build: bump `vercel-php` from 0.5.1 to 0.5.2
"Error: The Runtime @vercel/vc-build is using nodejs12.x, which was
discontinued by an upstream provider. In turn, we have to ask you to
upgrade your Runtime to a more recent version or consult the author for
more details. Learn More:
https://github.com/vercel/vercel/blob/master/DEVELOPING_A_RUNTIME.md#lambdaruntime"

Because vercel-php dropped node.js 12, so need to bump the project to
vercel-php@0.5.2 to use node js 14.x. Otherwise the project cannot be deployed to vercel.
link: https://vercel.com/changelog/node-js-12-is-being-deprecated
link: https://github.com/SCFOnedrive/SCFOnedrive.github.io/commit/01d970c94277ed481185ec5879919d7e4766a730

fix https://github.com/qkqpttgf/OneManager-php/issues/632

Signed-off-by: Hou Qiyao <shinka830@outlook.jp>